add simple install version of spruce ui installers;
